Precursor work for reordering layout optimization pass Contains work needed to do optimization pass that allows for nodes with different layout orders
Story #1622: Add identity specialization for reorder node
Story #1620: All conv. nodes should refine to have reshape nodes before and after node
Story #1619: Add input and output layout infra to nodes for optimization work

Related work items: #1619, #1620, #1622

(cherry picked from commit 05e8b5769973f5e31e42e5c4937df8976a453c27 (PR !907))

+++ libraries/nodes/src/MatrixMatrixMultiplyNode.cpp
@@ -22,7 +22,7 @@ namespace nodes
         // Relevant archive format versions
         //
         constexpr utilities::ArchiveVersion currentArchiveVersion = {utilities::ArchiveVersionNumbers::v2};
-        
+
         template <typename ValueType>
         void MatrixMatrixMultiply(bool transposeA, bool transposeB, bool transposeC, int m, int n, int k, const std::vector<ValueType>& matrixAValues, const std::vector<ValueType>& matrixBValues, std::vector<ValueType>& matrixCValues)
         {
@@ -174,8 +174,8 @@ namespace nodes
     template<typename ValueType>
     void MatrixMatrixMultiplyNode<ValueType>::Compute() const
     {
-        assert(input1.Size() == _m * _k);
-        assert(input2.Size() == _k * _n);
+        assert(static_cast<int>(input1.Size()) == _m * _k);
+        assert(static_cast<int>(input2.Size()) == _k * _n);
         auto inputMatrix1Values = input1.GetValue();
         auto inputMatrix2Values = input2.GetValue();
         std::vector<ValueType> outputMatrixValues(_m * _n);
@@ -204,7 +204,7 @@ namespace nodes
 
     template<typename ValueType>
     void MatrixMatrixMultiplyNode<ValueType>::Compile(model::IRMapCompiler& compiler, emitters::IRFunctionEmitter& function)
-    {  
+    {
         llvm::Value* pInput1 = compiler.EnsurePortEmitted(input1);
         llvm::Value* pInput2 = compiler.EnsurePortEmitted(input2);
         llvm::Value* pOutput = compiler.EnsurePortEmitted(output);
